+++ api/routes/confirm.js
@@ -1,8 +1,8 @@
 import express from 'express'
-import { confirmCode } from '../controllers/auth.js'
+import { verifyRegistrationCode } from '../controllers/auth.js'
 
 const router = express.Router()
 
-router.get('/token/:confirmationCode', confirmCode)
+router.get('/token/:confirmationCode', verifyRegistrationCode)
 
 export default router